Re-structured and improved tests
